fetch: 심화 1차
