+++ src/@lib/context/AppContext.tsx
@@ -1,0 +1,21 @@
+import { createContext, useState } from "react";
+import { useMemo, useCallback } from "../hooks";
+import { AppContextType } from "./types";
+
+export const AppContext = createContext<AppContextType | undefined>(undefined);
+
+export const AppContextProvider = ({ children }: { children: React.ReactNode }) => {
+  const [theme, setTheme] = useState<string>("light");
+
+  const toggleTheme = useCallback(() => {
+    setTheme((prev) => (prev === "light" ? "dark" : "light"));
+  }, [setTheme]);
+
+  const appValue = useMemo(
+    () => ({ theme, toggleTheme }),
+    [theme, toggleTheme],
+  );
+
+  return <AppContext.Provider value={appValue}>{children}</AppContext.Provider>;
+};
+
