+++ src/@lib/context/AuthContext.tsx
@@ -1,0 +1,34 @@
+import { createContext, useMemo, useState } from "react";
+import { AuthContextType, User } from "./types";
+import React from "react";
+import { useNotificationContext } from "./useNotifiCationContext";
+import { useCallback } from "../hooks";
+
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
+
+export const AuthContextProvider = ({children}: {children: React.ReactNode}) => {
+    const [user, setUser] = useState<User | null>(null);
+    const { addNotification } = useNotificationContext();
+    
+    const login = useCallback(
+        (email: string) => {
+          setUser({ id: 1, name: "홍길동", email });
+          addNotification("성공적으로 로그인되었습니다", "success");
+        },
+        [setUser, addNotification],
+      );
+    
+    const logout = useCallback(() => {
+        setUser(null);
+        addNotification("로그아웃되었습니다", "info");
+        }, [setUser, addNotification]);
+
+    const authValue = useMemo(
+        () => ({user, login, logout}),
+        [user, login, logout]
+    );
+
+    return (
+        <AuthContext.Provider value={authValue}>{children}</AuthContext.Provider>
+    );
+}
