+++ src/@lib/context/NotificationContext.tsx
@@ -1,0 +1,42 @@
+import { createContext, useState } from "react";
+import { NotificationContextType, Notification } from "./types";
+import { useCallback, useMemo } from "../hooks";
+
+export const NotificationContext = createContext<
+  NotificationContextType | undefined
+>(undefined);
+
+export const NotificationContextProvider = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const addNotification = useCallback(
+    (message: string, type: Notification["type"]) => {
+      setNotifications((prev) => [...prev, { id: Date.now(), message, type }]);
+    },
+    [setNotifications],
+  );
+
+  const removeNotification = useCallback(
+    (id: number) => {
+      setNotifications((prev) =>
+        prev.filter((notification) => notification.id !== id),
+      );
+    },
+    [setNotifications],
+  );
+
+  const notificationValue = useMemo(
+    () => ({ notifications, addNotification, removeNotification }),
+    [notifications, addNotification, removeNotification],
+  );
+
+  return (
+    <NotificationContext.Provider value={notificationValue}>
+      {children}
+    </NotificationContext.Provider>
+  );
+};
