+++ src/@lib/context/types.ts
@@ -21,9 +21,16 @@ export interface Notification {
 export interface AppContextType {
   theme: string;
   toggleTheme: () => void;
+}
+
+// AppContext 타입 정의
+export interface AuthContextType {
   user: User | null;
   login: (email: string, password: string) => void;
   logout: () => void;
+}
+
+export interface NotificationContextType {
   notifications: Notification[];
   addNotification: (message: string, type: Notification["type"]) => void;
   removeNotification: (id: number) => void;
