+++ src/@lib/context/useAuthContexts.ts
@@ -1,0 +1,13 @@
+import { createContext, useContext } from "react";
+import { AuthContextType } from "./types";
+
+const AuthContext = createContext<AuthContextType | undefined>(undefined);
+
+// 커스텀 훅: useAppContext
+export const useAuthContext = () => {
+  const context = useContext(AuthContext);
+  if (context === undefined) {
+    throw new Error("useAuthContext must be used within an AppProvider");
+  }
+  return context;
+};
